% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-lines.R
\name{as_lines}
\alias{as_lines}
\title{Print vector elements on multiple lines}
\usage{
as_lines(...)
}
\arguments{
\item{...}{Objects to print.}
}
\value{
\code{NULL} invisibly.
}
\description{
Thin wrapper around \code{\link[=cat]{cat()}} to display vector elements on
multiple lines when rendering an R Markdown or Quarto document. This is
primarily intended to be used with \code{\link{Plume}} methods to output each
returned element on its own line.
}
\examples{
aut <- Plume$new(encyclopedists)
as_lines(aut$get_affiliations())
}
