\name{Briggs}
\alias{Briggs}
\title{Briggs}
\usage{
  Briggs(stability = "D")
}
\arguments{
  \item{stability}{Pasquill stability class}
}
\value{
  a function sigma(x) returning a list with y and z
  components
}
\description{
  Briggs neutral curves for urban areas
}
\references{
  Briggs, G.A., 1973. Diffusion estimation for small
  emissions. ATDL contribution, File No. 79. Air Resources
  Atmospheric Turbulence and Diffusion Laboratory. NOAA,
  Oakridge, Tennessee, 59pp.
}

