% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate vector elements}
\usage{
enumerate(x, sep = ", ", last = " and ")
}
\arguments{
\item{x}{A character vector.}

\item{sep}{Separator used to separate the terms.}

\item{last}{Separator used to separate the last two items if \code{x} has at least
2 items.}
}
\value{
A character string with the same class as \code{x}.
}
\description{
Wrapper around \code{\link[glue:glue_collapse]{glue_collapse()}} using
\code{sep = ", "} and \code{last = " and "} as default arguments.
}
\examples{
aut <- Plume$new(encyclopedists)
aut$get_author_list() |> enumerate()
}
