% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name-handler.R
\name{NameHandler}
\alias{NameHandler}
\title{NameHandler class}
\description{
Internal class that handles the names of a \code{plume} object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NameHandler-new}{\code{NameHandler$new()}}
\item \href{#method-NameHandler-clone}{\code{NameHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-NameHandler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameHandler$new(names)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NameHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
