% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{orcid}
\alias{orcid}
\title{ORCID icon}
\usage{
orcid(size = 16, bw = FALSE)
}
\arguments{
\item{size}{Size of the icon (in pixels).}

\item{bw}{Should the black and white version of the icon be used?}
}
\value{
A plume icon.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=icn_orcid]{icn_orcid()}} instead.
}
\keyword{internal}
