% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-symbols.R
\name{plm_symbols}
\alias{plm_symbols}
\title{Set symbols for \code{Plume}}
\usage{
plm_symbols(
  affiliation = NULL,
  corresponding = "*",
  note = c("†", "‡", "§", "¶", "#", "**")
)
}
\arguments{
\item{affiliation, corresponding, note}{Character vectors of symbols to use,
or \code{NULL} to use numerals.}
}
\value{
A named list.
}
\description{
Set the symbols used in a \code{\link{Plume}} object.
}
\details{
Note that you should not escape special characters yourself when using this
function, as this is done internally by \code{\link{Plume}}.
}
\examples{
aut <- Plume$new(
  encyclopedists,
  symbols = plm_symbols(affiliation = sequential(letters))
)
}
