% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plume-package.R
\docType{package}
\name{plume-package}
\alias{plume}
\alias{plume-package}
\title{plume: A Simple Author Handler for Scientific Writing}
\description{
Handles and formats author information in scientific writing in 'R Markdown' and 'Quarto'. 'plume' provides easy-to-use and flexible tools for inserting author data in 'YAML' as well as generating author and contribution lists (among others) as strings from tabular data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arnaudgallou.github.io/plume/}
  \item \url{https://github.com/arnaudgallou/plume}
  \item Report bugs at \url{https://github.com/arnaudgallou/plume/issues}
}

}
\author{
\strong{Maintainer}: Arnaud Gallou \email{arangacas@gmail.com} (\href{https://orcid.org/0000-0002-1002-4247}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Arnaud Gallou \email{arangacas@gmail.com} (\href{https://orcid.org/0000-0002-1002-4247}{ORCID}) [copyright holder]
}

}
\keyword{internal}
