% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plume_template.R
\name{plume_template}
\alias{plume_template}
\title{Create a table template for plume classes}
\usage{
plume_template(minimal = TRUE)
}
\arguments{
\item{minimal}{If \code{TRUE}, returns an empty tibble with the following columns:
\code{given_name}, \code{family_name}, \code{email}, \code{orcid}, \code{affiliation}, \code{contribution}
and \code{note}. Otherwise the function returns a template with all columns that
can be supplied to plume classes.}
}
\value{
An empty tibble.
}
\description{
This helper function allows you to generate an empty
\code{\link[tibble:tibble]{tibble}} that you can use as a template to supply author
data.
}
