% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-setter.R
\name{StatusSetterPlume}
\alias{StatusSetterPlume}
\title{StatusSetterPlume class}
\description{
Internal class extending \code{StatusSetter} for \code{Plume}.
}
\keyword{internal}
\section{Super classes}{
\code{NameHandler} -> \code{\link[plume:PlumeHandler]{PlumeHandler}} -> \code{\link[plume:StatusSetter]{StatusSetter}} -> \code{StatusSetterPlume}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-StatusSetterPlume-set_main_contributors}{\code{StatusSetterPlume$set_main_contributors()}}
    \item \href{#method-StatusSetterPlume-clone}{\code{StatusSetterPlume$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="data"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-data'><code>PlumeHandler$data()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="get_plume"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-get_plume'><code>PlumeHandler$get_plume()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="get_roles"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-get_roles'><code>PlumeHandler$get_roles()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="print"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-print'><code>PlumeHandler$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="roles"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-roles'><code>PlumeHandler$roles()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="plume" data-topic="StatusSetter" data-id="initialize"><a href='../../plume/html/StatusSetter.html#method-StatusSetter-initialize'><code>StatusSetter$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="plume" data-topic="StatusSetter" data-id="set_corresponding_authors"><a href='../../plume/html/StatusSetter.html#method-StatusSetter-set_corresponding_authors'><code>StatusSetter$set_corresponding_authors()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetterPlume-set_main_contributors"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetterPlume-set_main_contributors}{}}}
\subsection{\code{StatusSetterPlume$set_main_contributors()}}{
  Force one or more contributors' names to appear first in the
contribution list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{StatusSetterPlume$set_main_contributors(..., .roles = NULL, .by = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{One or more unquoted expressions separated by commas.
Expressions matching values in the column defined by \code{by} or \code{.by}
determine main contributors. Expressions can be named after any role to
set different main contributors to different roles at once, in which
case the \code{.roles} parameter only applies roles that are not already set
to unnamed expressions. Matching of values is case-insensitive and dot-
agnostic.}
      \item{\code{.roles}}{Roles to assign main contributors to. If \code{.roles} is a
named vector, only the names will be used.}
      \item{\code{.by}}{Variable used to specify which authors are main contributors.
By default, uses authors' id.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The class instance.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetterPlume-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetterPlume-clone}{}}}
\subsection{\code{StatusSetterPlume$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{StatusSetterPlume$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
