% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-setter.R
\name{StatusSetterQuarto}
\alias{StatusSetterQuarto}
\title{StatusSetterQuarto class}
\description{
Internal class extending \code{StatusSetter} for \code{PlumeQuarto}.
}
\section{Super classes}{
\code{\link[plume:NameHandler]{plume::NameHandler}} -> \code{\link[plume:PlumeHandler]{plume::PlumeHandler}} -> \code{\link[plume:StatusSetter]{plume::StatusSetter}} -> \code{StatusSetterQuarto}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatusSetterQuarto-new}{\code{StatusSetterQuarto$new()}}
\item \href{#method-StatusSetterQuarto-set_equal_contributor}{\code{StatusSetterQuarto$set_equal_contributor()}}
\item \href{#method-StatusSetterQuarto-set_deceased}{\code{StatusSetterQuarto$set_deceased()}}
\item \href{#method-StatusSetterQuarto-clone}{\code{StatusSetterQuarto$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="get_plume"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-get_plume'><code>plume::PlumeHandler$get_plume()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="print"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-print'><code>plume::PlumeHandler$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plume" data-topic="StatusSetter" data-id="set_corresponding_authors"><a href='../../plume/html/StatusSetter.html#method-StatusSetter-set_corresponding_authors'><code>plume::StatusSetter$set_corresponding_authors()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetterQuarto-new"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetterQuarto-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusSetterQuarto$new(..., by)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetterQuarto-set_equal_contributor"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetterQuarto-set_equal_contributor}{}}}
\subsection{Method \code{set_equal_contributor()}}{
Set equal contributors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusSetterQuarto$set_equal_contributor(..., by)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more unquoted expressions separated by commas.
Expressions matching values in the column defined by \code{by} determine
equal contributors. Matching of values is case-insensitive.}

\item{\code{by}}{Variable used to specify which authors are equal contributors.
By default, uses authors' id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The class instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetterQuarto-set_deceased"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetterQuarto-set_deceased}{}}}
\subsection{Method \code{set_deceased()}}{
Set deceased authors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusSetterQuarto$set_deceased(..., by)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more unquoted expressions separated by commas.
Expressions matching values in the column defined by \code{by} determine
deceased authors. Matching of values is case-insensitive.}

\item{\code{by}}{Variable used to specify whether an author is deceased or not.
By default, uses authors' id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The class instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetterQuarto-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetterQuarto-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusSetterQuarto$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
