% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-lines.R
\name{as_lines}
\alias{as_lines}
\title{Print vector elements on distinct lines}
\usage{
as_lines(...)
}
\arguments{
\item{...}{Objects to print.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Thin wrapper around \code{cat()} to display vector elements on distinct lines when
rendering an R Markdown or Quarto document.
}
\examples{
aut <- Plume$new(encyclopedists)
as_lines(aut$get_affiliations())
}
