% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_roles.R
\name{credit_roles}
\alias{credit_roles}
\title{CRediT roles}
\usage{
credit_roles(oxford_spelling = TRUE)
}
\arguments{
\item{oxford_spelling}{Should the suffix -ize/-ization be used?}
}
\value{
A named vector.
}
\description{
Helper function returning the 14 contributor roles of the
\href{https://credit.niso.org}{Contributor Roles Taxonomy} (CRediT). This function is the default argument of the
\code{roles} and \code{role_cols} parameters in plume classes and \code{\link[=plm_template]{plm_template()}},
respectively.
}
\examples{
credit_roles()
}
