% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-encyclopedists.R
\docType{data}
\name{encyclopedists}
\alias{encyclopedists}
\alias{encyclopedists_fr}
\title{Famous encyclopedists}
\format{
A tibble with 4 rows and 12 variables:
\describe{
\item{given_name,prénom}{authors' given names}
\item{family_name,nom}{authors' family names}
\item{email,courriel}{authors' email addresses}
\item{phone,téléphone}{authors' phone numbers}
\item{orcid}{authors' ORCID}
\item{affiliation}{authors' affiliations}
\item{role_n}{authors' roles written as nouns}
\item{role_v}{authors' roles written as active verbs}
\item{note}{special notes about authors}
}
}
\usage{
encyclopedists

encyclopedists_fr
}
\description{
Data on four famous authors of the Encyclopédie (originally "Encyclopédie, ou
dictionnaire raisonné des sciences, des arts et des métiers") published in
France in the second half of the 18th century. The data set is available in
English (\code{encyclopedists}) and French (\code{encyclopedists_fr}).
}
\examples{
encyclopedists

encyclopedists_fr
}
\keyword{datasets}
