% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-setter.R
\name{everyone}
\alias{everyone}
\alias{everyone_but}
\title{Plume selectors}
\usage{
everyone()

everyone_but(...)
}
\arguments{
\item{...}{One or more unquoted expressions separated by commas. Expressions
matching values in the column defined by the \code{by} parameter of \verb{set_*()}
methods are used to set a given status to authors. Matching of values is
case-insensitive.}
}
\description{
Helper functions used to assign status to multiple authors at
once.
}
\examples{
aut <- Plume$new(encyclopedists)
aut$set_corresponding_authors(everyone())

aut$set_corresponding_authors(everyone_but(jean), by = "given_name")
}
