% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{orcid}
\alias{orcid}
\title{ORCID icon}
\usage{
orcid(size = 16, bw = FALSE)
}
\arguments{
\item{size}{Size (in pixels) of the icon.}

\item{bw}{Should the black and white version of the icon be used?}
}
\value{
A plume icon.
}
\description{
Helper function to control the size and colour of the ORCID
icon.
}
\examples{
aut <- Plume$new(encyclopedists, orcid_icon = orcid(bw = TRUE))
}
