% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-template.R
\name{plm_template}
\alias{plm_template}
\title{Create a table template for plume classes}
\usage{
plm_template(minimal = TRUE, credit_roles = FALSE)
}
\arguments{
\item{minimal}{If \code{TRUE}, returns an empty tibble with the following columns:
\code{given_name}, \code{family_name}, \code{email}, \code{orcid}, \code{affiliation}, \code{role} and
\code{note}. Otherwise the function returns a template with all columns that can
be supplied to plume classes that are not \code{PlumeQuarto}-specific.}

\item{credit_roles}{Should the \href{https://credit.niso.org}{Contributor Roles Taxonomy} be used?}
}
\value{
An empty tibble.
}
\description{
This helper function allows you to generate an empty
\code{\link[tibble:tibble]{tibble}} that you can use as a template to supply
author data.
}
\examples{
plm_template()
}
