% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plume-package.R
\docType{package}
\name{plume-package}
\alias{plume}
\alias{plume-package}
\title{plume: A Simple Author Handler for Scientific Writing}
\description{
Tools to reduce the hassle of handling and formatting author
information in scientific writing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arnaudgallou/plume}
  \item \url{https://arnaudgallou.github.io/plume/}
  \item Report bugs at \url{https://github.com/arnaudgallou/plume/issues}
}

}
\author{
\strong{Maintainer}: Arnaud Gallou \email{arangacas@gmail.com} (\href{https://orcid.org/0000-0002-1002-4247}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Hadley Wickham (Author of included stringb functions.) [copyright holder]
}

}
\keyword{internal}
