% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plume-package.R
\docType{package}
\name{plume-package}
\alias{plume}
\alias{plume-package}
\title{plume: An Author List Handler}
\description{
A tool for handling and printing author information in Rmarkdown
and Quarto.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arnaudgallou/plume}
  \item \url{https://arnaudgallou.github.io/plume/}
  \item Report bugs at \url{https://github.com/arnaudgallou/plume/issues}
}

}
\author{
\strong{Maintainer}: Arnaud Gallou \email{agallou.gh@gmail.com} (\href{https://orcid.org/0000-0002-1002-4247}{ORCID})

}
\keyword{internal}
