\docType{package}
\name{plume}
\alias{plume}
\alias{plume-package}
\title{plume}
\description{
  Demo framework for Gaussian dispersion modeling
}
\author{
  David Holstius \email{david.holstius@berkeley.edu}
}
\references{
  Benson, P. (1979) CALINE-3: A versatile dispersion model
  for predicting air pollution levels near highways and
  urban streets. Federal Highway Authority report
  FHWA/CA/TL-79/23, California DOT, Sacramento (1979).
}
\keyword{atmospheric}
\keyword{disperson}
\keyword{environmental}
\keyword{modeling}
\keyword{package}
\keyword{plume}

