% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plume-handler.R
\name{PlumeHandler}
\alias{PlumeHandler}
\title{PlumeHandler class}
\description{
Internal class processing and shaping tabular data into a
\code{plume} object.
}
\section{Super class}{
\code{\link[plume:NameHandler]{plume::NameHandler}} -> \code{PlumeHandler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlumeHandler-new}{\code{PlumeHandler$new()}}
\item \href{#method-PlumeHandler-print}{\code{PlumeHandler$print()}}
\item \href{#method-PlumeHandler-get_plume}{\code{PlumeHandler$get_plume()}}
\item \href{#method-PlumeHandler-get_roles}{\code{PlumeHandler$get_roles()}}
\item \href{#method-PlumeHandler-clone}{\code{PlumeHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlumeHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlumeHandler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumeHandler$new(
  data,
  names,
  roles,
  credit_roles,
  initials_given_name,
  family_name_first = FALSE,
  interword_spacing = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlumeHandler-print"></a>}}
\if{latex}{\out{\hypertarget{method-PlumeHandler-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumeHandler$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlumeHandler-get_plume"></a>}}
\if{latex}{\out{\hypertarget{method-PlumeHandler-get_plume}{}}}
\subsection{Method \code{get_plume()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumeHandler$get_plume()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlumeHandler-get_roles"></a>}}
\if{latex}{\out{\hypertarget{method-PlumeHandler-get_roles}{}}}
\subsection{Method \code{get_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumeHandler$get_roles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlumeHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlumeHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumeHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
