% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_names.R
\name{set_default_names}
\alias{set_default_names}
\title{Set new default names to a plume subclass}
\usage{
set_default_names(...)
}
\arguments{
\item{...}{Key-value pairs where keys are old names and values their respective
replacements.}
}
\value{
A named list.
}
\description{
This helper function allows you to set new default names to a plume
subclass, e.g. to set default names to a language other than English. See
\code{vignette("working-in-other-languages")} for more details.
}
\details{
Available names are:

\code{id}, \code{initials}, \code{literal_name}, \code{corresponding}, \code{deceased}, \code{equal_contributor}, \code{given_name}, \code{family_name}, \code{dropping_particle}, \code{email}, \code{orcid}, \code{number}, \code{fax}, \code{url}, \code{affiliation}, \code{contribution}, \code{note}
}
