% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-setter.R
\name{StatusSetter}
\alias{StatusSetter}
\title{StatusSetter class}
\description{
Class setting authors' status
}
\section{Super classes}{
\code{\link[plume:NameHandler]{plume::NameHandler}} -> \code{\link[plume:PlumeHandler]{plume::PlumeHandler}} -> \code{StatusSetter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatusSetter-set_corresponding_authors}{\code{StatusSetter$set_corresponding_authors()}}
\item \href{#method-StatusSetter-clone}{\code{StatusSetter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="get_plume"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-get_plume'><code>plume::PlumeHandler$get_plume()</code></a></li>
<li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="initialize"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-initialize'><code>plume::PlumeHandler$initialize()</code></a></li>
<li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="print"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-print'><code>plume::PlumeHandler$print()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetter-set_corresponding_authors"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetter-set_corresponding_authors}{}}}
\subsection{Method \code{set_corresponding_authors()}}{
Set corresponding authors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusSetter$set_corresponding_authors(..., by)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Values in the column defined by \code{by} used to specify
corresponding authors. Matching of values is case-insensitive. Use
\code{"all"} to assign \code{TRUE} to all authors.}

\item{\code{by}}{Variable used to set corresponding authors. By default, uses
authors' id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The class instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusSetter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
