% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-setter.R
\name{StatusSetterPlume}
\alias{StatusSetterPlume}
\title{StatusSetterPlume class}
\description{
Internal class extending \code{StatusSetter} for \code{Plume}.
}
\section{Super classes}{
\code{\link[plume:NameHandler]{plume::NameHandler}} -> \code{\link[plume:PlumeHandler]{plume::PlumeHandler}} -> \code{\link[plume:StatusSetter]{plume::StatusSetter}} -> \code{StatusSetterPlume}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatusSetterPlume-set_main_contributors}{\code{StatusSetterPlume$set_main_contributors()}}
\item \href{#method-StatusSetterPlume-clone}{\code{StatusSetterPlume$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="get_plume"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-get_plume'><code>plume::PlumeHandler$get_plume()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plume" data-topic="PlumeHandler" data-id="print"><a href='../../plume/html/PlumeHandler.html#method-PlumeHandler-print'><code>plume::PlumeHandler$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plume" data-topic="StatusSetter" data-id="initialize"><a href='../../plume/html/StatusSetter.html#method-StatusSetter-initialize'><code>plume::StatusSetter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plume" data-topic="StatusSetter" data-id="set_corresponding_authors"><a href='../../plume/html/StatusSetter.html#method-StatusSetter-set_corresponding_authors'><code>plume::StatusSetter$set_corresponding_authors()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetterPlume-set_main_contributors"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetterPlume-set_main_contributors}{}}}
\subsection{Method \code{set_main_contributors()}}{
Set main contributors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusSetterPlume$set_main_contributors(..., .roles = NULL, .by)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more unquoted expressions separated by commas.
Expressions matching values in the column defined by \code{.by} determine
main contributors. Matching of values is case-insensitive.
Alternatively, you can provide a list of key-value pairs where keys
define roles and values identify contributors.}

\item{\code{.roles}}{Roles to assign main contributors to.}

\item{\code{.by}}{Variable used to specify which authors are equal contributors.
By default, uses authors' id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The class instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatusSetterPlume-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatusSetterPlume-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusSetterPlume$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
